broadcast: merge pure helpers into one lib/broadcast-utils module
The pure broadcast helpers were split across src/lib/player/broadcast-utils.js
and src/lib/player/broadcast-payload.js, under the player directory despite
being broadcast-domain, with the test file stranded next to broadcast.js.

- merge both into src/lib/broadcast-utils.js, sibling of broadcast.js
- drop the unused calculateSeekTime/DRIFT_TOLERANCE_SECONDS re-export
- rename broadcast.test.js to broadcast-utils.test.js to match what it tests
- drive pickBroadcastFields off BROADCAST_STATE_DEFAULTS keys so the
  field list lives in one place
- update reference.json, broadcast.md and plan.md
